requirements.txt
discord.py
openai